refactored unit tests to separate address and transaction and adjust sign_raw to match new return type
