+++ .github/workflows/ci.yml
@@ -205,7 +205,7 @@ jobs:
               "x86_64-linux-dbg") 
               make check -j"$(getconf _NPROCESSORS_ONLN)" V=1
               python3 tooltests.py
-              ./bindings/py_wrappers/pytest/cython_tests.sh
+              ./bindings/python/pytest/cython_tests.sh
               ;;
               "i686-linux") 
               make check -j"$(getconf _NPROCESSORS_ONLN)" V=1
