+++ bindings/golang/libdogecoin/go.mod
@@ -1,0 +1,3 @@
+module libdogecoin
+
+go 1.18
